+++ tests/testthat/test_dplyr_verbs.R
@@ -41,16 +41,12 @@ test_that('drop_na works',{
 })
 
 test_that("summarize `.groups` argument matches dplyr behavior (0 groups case)", {
-  lapply(c("keep", "drop_last", "drop"), function(group_type) {
+  lapply(c("keep", "drop_last", "drop", "rowwise"), function(group_type) {
     expect_equal(
       dstrata %>% summarize(x = unweighted(n()), .groups = !!group_type),
       as_tibble(apistrat) %>% summarize(x = n(), .groups = !!group_type)
     )
   })
-
-  # rowwise doesn't match dplyr's behavior for 0 groups. for now assuming this will be changed in dplyr
-  # https://github.com/tidyverse/dplyr/issues/5422
-  expect_is(dstrata %>% summarize(x = unweighted(n()), .groups = "rowwise"), "rowwise_df")
 })
 
 test_that("summarize `.groups` argument matches dplyr behavior (1 groups case)", {
@@ -118,19 +114,6 @@ test_that("ungrouped summarize accepts mix of 1 row & multi row results", {
   expect_equal(direct, round_about)
 })
 
-test_that("ungrouped summarize errors when multi row results of different number of rows", {
-  expect_error(
-    direct <- dstrata %>%
-      summarize(
-        w = survey_mean(api99),
-        x = unweighted(quantile(api99, c(0.05, 0.5, 0.95))),
-        y = survey_mean(api00),
-        z = unweighted(quantile(api00, c(0.05, 0.5, 0.75, 0.95)))
-      ),
-    "summarise results for argument `z` must be size 1 or 3 but it is 4"
-  )
-})
-
 test_that("grouped summarize accepts mix of 1 row & multi row results", {
   direct <- dstrata %>%
     group_by(both) %>%
@@ -179,16 +162,13 @@ test_that("grouped summarize accepts mix of 1 row & multi row results", {
   expect_equal(direct, round_about)
 })
 
-test_that("grouped summarize errors when multi row results of different number of rows", {
-  expect_error(
-    dstrata %>%
-      group_by(both) %>%
-      summarize(
-        w = survey_mean(api99),
-        x = unweighted(quantile(api99, c(0.05, 0.5, 0.95))),
-        y = survey_mean(api00),
-        z = unweighted(quantile(api00, c(0.05, 0.5, 0.75, 0.95)))
-      ),
-    "summarise results for argument `z` must be size 1 or 3 but it is 4 for group: both = 1"
-  )
+test_that("group_trim works", {
+  before_trim <- dstrata %>%
+    group_by(both, .drop = FALSE) %>%
+    filter(both == "No")
+
+  expect_equal(group_keys(before_trim)[[1]], factor(c("No", "Yes"), c("No", "Yes")))
+
+  after_trim <- before_trim %>% group_trim()
+  expect_equal(group_keys(after_trim)[[1]], factor(c("No"), c("No")))
 })